Helper: Reject reports with missing extension
Currently the Helper is supposed to abort an aggregation job that
contains a report share that is missing the `taskprov` extension.
However because the Helper report shares are encrypted, the Leader has
no way of knowing whether the extension is present or not. This means
there is no protocol-level mechanism for the Helper to convey which
report needs to be removed from the candidate set before retrying.

Instead of aborting, have the Helper simply reject the malformed report.

Clarify a few things while at it.