Flesh out IANA considerations and extending taskbind (*)
* Ask IANA to register the extension codepoint
* Ask IANA to create a registry for DP mechanisms
* Add guidance for extending "taskbind" in the future
* Ask IANA to add "invalidTask" to the URN sub-namespace for DAP

We haven't quite figured out the last of these, so leave it as a TODO.

Future documents might extend the batch config, DP config, or VDAF
config, so it's helpful if these structures are self-contained. To
accomplish this, move the time precision and min batch size parameters
out of BatchConfig into TaskConfig and move the DP mechnaism and config
out of VdafConfig.