Improve extension points for batch modes and VDAFs (*)
Along the same lines as
https://github.com/ietf-wg-ppm/draft-ietf-ppm-dap/pull/627

Harmonize the extension points for new batch modes or VDAFs with
Tasbind extensions by following the Tag-Length-Value (TLV) pattern. That
is, encode the variant followed by length-prefixed payload, the format
of which is determined by the variant. This is inteded to make it easier
to extend the document in the future.

This change also has a couple of bug fixes:

* Fix encoding of Prio3Sum config (the circuit now requires a maximum
  measurement rather than the bit length of the maximum measurement).
* Add encoding for Prio3MultihotCountVec