Add TimePrecision type and redefine time_precision as it
DAP-16 redefined Time, Duration (and by extension Interval) in terms
of multiples of the task's time_precision, making `Duration time_precision`
nonsensical in TaskProv. Since we do want to keep and use the new `Time`
and `Duration` definitions here, this adds a new `TimePrecision` type and
defines it simply, so the other types can depend on it.

I marked this change as breaking, though it really isn't -- but the impacts
from DAP-16 _are_, which I think this change effectuates.

Bikeshed: I considered calling the type `Seconds`, but that feels like
generality for generality's sake, while this is deliberately specific.

Fixes #110